new uppdate with static methods

+++ src/main/java/dao/InterfaseDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface InterfaseDao<T> {
     public void add(T t) throws SQLException;
     public void delete(T t) throws SQLException;
-    public T get(int id) throws SQLException;
+    public T get(long id) throws SQLException;
     public List<T> getAll() throws SQLException;
     public void update(T t) throws SQLException;
 }
